Vegetation Placement
The greenery control implemented here took a while of tuning to get
good default parameters. Random noise is scattered across the map,
biased by height, and then a first pass adds weights for being near
water before a final pass smooths using a parameter based on the size
of the map (but which scales sanely).

Also, a nice new colouring function, life, which brings it out.